feat:const refactoring
