+++ src/constants/propTypes.js
@@ -8,7 +8,7 @@ import PropTypes from 'prop-types';
 export const productPropType = PropTypes.shape({
   _id: PropTypes.string.isRequired,
   name: PropTypes.string.isRequired,
-  type: PropTypes.oneOf(['s', 'main', 'sauce']).isRequired,
+  type: PropTypes.oneOf(['bun', 'main', 'sauce']).isRequired,
   proteins: PropTypes.number.isRequired,
   fat: PropTypes.number.isRequired,
   carbohydrates: PropTypes.number.isRequired,
